requirements.txt
streamlit>=1.38
Pillow>=10.3
numpy>=1.26
